Try to deal with some circular issues with the logging module
